requirements.txt
python-dotenv
pandas
numpy
scikit-learn
pymongo
#pymongo[srv]==3.6
certifi
dill
pyaml
# -e .